minor restructuring
