--- src/components/Hooks/index.js
@@ -1,3 +1,0 @@
-export * from './useEventListener'
-export * from './useOnClickOutside'
-export * from './useLegacyState'
