+++ src/hooks/useExecom.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { useStaticQuery, graphql } from "gatsby";
 
-export const useExecom = (chair=false) => {
-    const { allMembers, allChairs } = useStaticQuery(
+export const useExecom = () => {
+    const { allChairs } = useStaticQuery(
         graphql`
             query {
                 allChairs: allExecomMembersJson(sort: {fields: year, order: DESC}, limit: 1) {
@@ -29,9 +29,7 @@ export const useExecom = (chair=false) => {
              }
         `
     )
-    if(!chair){
-        return allMembers;
-    }
+
     return allChairs;
 }
