+++ src/pages/about.jsx
@@ -2,9 +2,9 @@ import React  from "react";
 import Helmet from "react-helmet";
 import About from "../components/About/About";
 import config from "../../data/SiteConfig";
-import Sidebar from "../components/Sidebar/Sidebar";
+import Sidebar from "../components/sidebar";
 import { Row, Col } from "react-bootstrap";
-import Breadcrumb from "../components/Breadcrumbs";
+import Breadcrumb from "../components/breadcrumbs";
 export default ({ pageContext }) => {
     const {
         breadcrumb: { crumbs },
